fix image remove form server

+++ script/deleteImage.php
@@ -11,10 +11,14 @@ if (isset($_SESSION['user'])) {
     if (isset($_POST['imageId'])) {
         if (($image = Image::createFromId($_POST['imageId'])) !== false) {
             if ($image->getUserId() === $_SESSION['user']['userId']) {
-                $image->delete();
-                $json['message'] = "Image removed";
-                $json['valid'] = true;
-                $json['userId'] = $_SESSION['user']['userId'];
+                $img_path = "../" . $image->getPath();  
+                if (unlink($img_path)) {
+                    $image->delete();
+                    $json['message'] = "Image removed";
+                    $json['valid'] = true;
+                    $json['userId'] = $_SESSION['user']['userId'];
+                } else
+                    $json['message'] = "Failed to remove image";
             } else
                 $json['message'] = "You don't have the permission to delete this image";
         }
